Refactors sticker metadata logic to dedicated module
Improves modularity and code organization by extracting the `addStickerMetadata` function into its own file. This separates the concern of handling sticker EXIF data from the primary sticker command processing.
